Move Icon from core to icons package
