+++ packages/core/src/components/_index.scss
@@ -13,7 +13,6 @@
 @import "editable-text/editable-text";
 @import "forms/index";
 @import "hotkeys/hotkeys";
-@import "icon/icon";
 @import "menu/menu";
 @import "navbar/navbar";
 @import "non-ideal-state/non-ideal-state";
