+++ packages/icons/src/blueprint-icons.scss
@@ -7,3 +7,4 @@ Licensed under the terms of the LICENSE file distributed with this project.
 
 @import "font-imports";
 @import "icons";
+@import "components/index";
