+++ packages/icons/src/components/_index.scss
@@ -1,0 +1,4 @@
+// Copyright 2015 Palantir Technologies, Inc. All rights reserved.
+// Licensed under the terms of the LICENSE file distributed with this project.
+
+@import "icon/icon";
