+++ packages/icons/src/components/icon/_icon.scss
