+++ packages/icons/src/components/index.ts
@@ -1,0 +1,7 @@
+/*
+ * Copyright 2018 Palantir Technologies, Inc. All rights reserved.
+ *
+ * Licensed under the terms of the LICENSE file distributed with this project.
+ */
+
+export * from "./icon/icon";
